refactor: rename root component to index.ts
