+++ packages/react-ui/src/app/builder/step-settings/loops-settings.tsx
@@ -5,7 +5,7 @@ import { ApMarkdown } from '@/components/custom/markdown';
 import { FormField, FormItem, FormLabel } from '@/components/ui/form';
 import { LoopOnItemsAction } from '@activepieces/shared';
 
-import { TextInputWithMentions } from '../piece-properties/text-input-with-mentions/text-input-with-mentions';
+import { TextInputWithMentions } from '../piece-properties/text-input-with-mentions';
 
 const markdown = `
 Select the items to iterate over from the previous step by clicking on the **Items** input, which should be a **list** of items.
